refactor(auth): consolidate OAuth buttons into OAuthLoginButton(s), drop SocialLoginButton
